# src/public-website/contracts/openapi/admin-api.yaml
openapi: 3.1.0
info:
  title: International Center Admin API
  description: Administrative API for International Center content management system
  version: 1.0.0
  contact:
    name: International Center API Team
servers:
  - url: https://admin-api.axiomcloud.dev/api/v1
    description: Production admin server
  - url: http://localhost:4001/api/v1
    description: Development admin server
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /news:
    get:
      summary: Get all news articles (admin)
      operationId: getNewsAdmin
      tags:
        - News Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          description: List of news articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsArticle'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      summary: Create new news article
      operationId: createNewsArticle
      tags:
        - News Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewsArticleRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'
...
  /news/categories:
    get:
      summary: Get all news categories (admin)
      operationId: getNewsCategoriesAdmin
      tags:
        - News Management
      responses:
        '200':
          description: News categories
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsCategory'

    post:
      summary: Create news category
      operationId: createNewsCategory
      tags:
        - News Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNewsCategoryRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '403':
          $ref: '#/components/responses/ErrorResponse'

  # Research management endpoints (similar pattern to news)
  /research:
    get:
      summary: Get all research publications (admin)
      operationId: getResearchAdmin
      tags:
        - Research Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          description: List of research publications
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ResearchPublication'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      summary: Create new research publication
      operationId: createResearchPublication
      tags:
        - Research Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResearchPublicationRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'

  # Services management endpoints
  /services:
    get:
      summary: Get all services (admin)
      operationId: getServicesAdmin
      tags:
        - Services Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      summary: Create new service
      operationId: createService
      tags:
        - Services Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'

  # Events management endpoints
  /events:
    get:
      summary: Get all events (admin)
      operationId: getEventsAdmin
      tags:
        - Events Management
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/StatusParam'
        - $ref: '#/components/parameters/CategoryIdParam'
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'

    post:
      summary: Create new event
      operationId: createEvent
      tags:
        - Events Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedResponse'

  # Inquiries management endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      $ref: './components/parameters/pagination.yaml#/PageParam'
    LimitParam:
      $ref: './components/parameters/pagination.yaml#/LimitParam'
    SearchParam:
      name: search
      in: query
      description: Search query string for full-text search
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: "cancer research"
    StatusParam:
      $ref: './components/parameters/filters.yaml#/StatusParam'
    CategoryIdParam:
      $ref: './components/parameters/filters.yaml#/CategoryIdParam'
  schemas:
...
    PaginationInfo:
      $ref: './public-api.yaml#/components/schemas/PaginationInfo'

    # Reference existing domain schemas
    Service:
      $ref: './components/schemas/services.yaml#/Service'
    NewsArticle:
      $ref: './components/schemas/news.yaml#/NewsArticle'
    NewsCategory:
      $ref: './components/schemas/news.yaml#/NewsCategory'
    ResearchPublication:
      $ref: './components/schemas/research.yaml#/ResearchPublication'
    Event:
      $ref: './components/schemas/events.yaml#/Event'

    # Admin-specific schemas
...
    CreateNewsArticleRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        summary:
          type: string
        content:
          type: string
          nullable: true
        category_id:
          type: string
          format: uuid
        image_url:
          type: string
          format: uri
          nullable: true
        author_name:
          type: string
          nullable: true
        news_type:
          type: string
          enum: [announcement, press_release, event, update, alert, feature]
        priority_level:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        tags:
          type: array
          items:
            type: string
        publishing_status:
          type: string
          enum: [draft, published, archived]
          default: draft
      required:
        - title
        - summary
        - category_id
        - news_type
...
    CreateNewsCategoryRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        description:
          type: string
          nullable: true
      required:
        - name
    CreateResearchPublicationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        abstract:
          type: string
        category_id:
          type: string
          format: uuid
        authors:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              affiliation:
                type: string
                nullable: true
              email:
                type: string
                format: email
                nullable: true
            required:
              - name
        research_type:
          type: string
          enum: [clinical_study, observational_study, review, case_study, meta_analysis, guideline, white_paper]
        study_status:
          type: string
          enum: [planned, ongoing, completed, published, archived]
        keywords:
          type: array
          items:
            type: string
      required:
        - title
        - abstract
        - category_id
        - authors
        - research_type
        - study_status
    CreateServiceRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        category_id:
          type: string
          format: uuid
        service_type:
          type: string
          enum: [consultation, treatment, procedure, diagnostic, therapy, rehabilitation, prevention]
        duration_minutes:
          type: integer
          minimum: 0
          nullable: true
        price_range:
          type: string
          nullable: true
        availability_status:
          type: string
          enum: [available, limited, unavailable, appointment_only]
        insurance_accepted:
          type: boolean
        telehealth_available:
          type: boolean
      required:
        - title
        - description
        - category_id
        - service_type
        - availability_status
        - insurance_accepted
        - telehealth_available
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        category_id:
          type: string
          format: uuid
        event_type:
          type: string
          enum: [conference, workshop, seminar, webinar, training, networking, fundraising, community]
        start_datetime:
          type: string
          format: date-time
        end_datetime:
          type: string
          format: date-time
          nullable: true
        registration_required:
          type: boolean
        max_capacity:
          type: integer
          minimum: 1
          nullable: true
        organizer:
          type: object
          properties:
            name:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
              nullable: true
          required:
            - name
            - email
      required:
        - title
        - description
        - category_id
        - event_type
        - start_datetime
        - registration_required
        - organizer
...
  responses:
...
    CreatedResponse:
      $ref: './components/responses/success.yaml#/CreatedResponse'
...
    ErrorResponse:
      $ref: './components/responses/errors.yaml#/ErrorResponse'
...